+++ app/vmui/packages/vmui/src/pages/QueryPage/HitsChart/hooks/useHitsChartConfig.ts
@@ -1,11 +1,11 @@
 import { useSearchParams } from "react-router-dom";
 import { useCallback, useMemo } from "preact/compat";
-import { LOGS_BAR_COUNT_DEFAULT, LOGS_LIMIT_HITS, WITHOUT_GROUPING } from "../../../../constants/logs";
+import { LOGS_LIMIT_HITS, WITHOUT_GROUPING } from "../../../../constants/logs";
 
 enum  HITS_PARAMS {
   TOP = "top_hits",
   GROUP = "group_hits",
-  BARS_COUNT = "bars_count",
+  STEP = "step",
 }
 
 export const useHitsChartConfig = () => {
@@ -16,24 +16,22 @@ export const useHitsChartConfig = () => {
     return Number.isFinite(n) && n > 0 ? n : LOGS_LIMIT_HITS;
   }, [searchParams]);
 
-  const barsCount = useMemo(() => {
-    const n = Number(searchParams.get(HITS_PARAMS.BARS_COUNT));
-    return Number.isFinite(n) && n > 0 ? n : LOGS_BAR_COUNT_DEFAULT;
+  const step = useMemo(() => {
+    return searchParams.get(HITS_PARAMS.STEP);
   }, [searchParams]);
 
   const groupFieldHits = searchParams.get(HITS_PARAMS.GROUP) || WITHOUT_GROUPING;
 
   const setValue = useCallback((param: HITS_PARAMS, newValue?: string | number) => {
     setSearchParams(prev => {
+      const prevValue = prev.get(param);
+
+      const nextValue = String(newValue);
+      const isEqual = nextValue === prevValue;
+      if (isEqual) return prev;
+
       const next = new URLSearchParams(prev);
-      const currentValue = prev.get(param);
-
-      if (newValue && newValue !== currentValue) {
-        next.set(param, String(newValue));
-      } else {
-        next.delete(param);
-      }
-
+      !nextValue ? next.delete(param) : next.set(param, nextValue);
       return next;
     });
   }, [setSearchParams]);
@@ -46,13 +44,13 @@ export const useHitsChartConfig = () => {
     setValue(HITS_PARAMS.GROUP, value);
   }, [setValue]);
 
-  const setBarsCount = useCallback((value?: number) => {
-    setValue(HITS_PARAMS.BARS_COUNT, value);
+  const setStep = useCallback((value?: string) => {
+    setValue(HITS_PARAMS.STEP, value);
   }, [setValue]);
 
   return {
     topHits: { value: topHits, set: setTopHits },
     groupFieldHits: { value: groupFieldHits, set: setGroupFieldHits },
-    barsCount: { value: barsCount, set: setBarsCount },
+    step: { value: step, set: setStep },
   };
 };